Move StockComparator, SharePriceComparator, and StockDataLoader into the use cases folder because they belong here.
