+++ src/main/java/use_cases/compare_stocks/StockComparator.java
@@ -1,4 +1,6 @@
-package main.java.entities;
+package main.java.use_cases.compare_stocks;
+
+import main.java.entities.Stock;
 
 public interface StockComparator {
     int compare(Stock stock1, Stock stock2);
